# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API для управления складом товаров с ролевой моделью доступа",
    "title": "MarketEase API",
    "contact": {},
    "version": "2.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
...
    "/products": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Получить список всех товаров с возможностью фильтрации и сортировки",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Получить все товары",
        "parameters": [
          {
            "type": "string",
            "description": "Фильтр по названию",
            "name": "name",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Минимальная цена",
            "name": "minPrice",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Максимальная цена",
            "name": "maxPrice",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Минимальное количество на складе",
            "name": "minStock",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Максимальное количество на складе",
            "name": "maxStock",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Поле для сортировки (name, price, stock, created_at)",
            "name": "sortBy",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Порядок сортировки (asc, desc)",
            "name": "sortOrder",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Включать удаленные товары",
            "name": "includeDeleted",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Product"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Добавить новый товар (требуется роль manager или director)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Создать новый товар",
        "parameters": [
          {
            "description": "Данные товара",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Product"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/products/deleted": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Получить список всех удаленных товаров (требуется роль manager или director)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Получить список удаленных товаров",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Product"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Получить информацию о товаре по его ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Получить товар по ID",
        "parameters": [
          {
            "type": "string",
            "description": "ID товара",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Product"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Изменить информацию о товаре (требуется роль manager или director)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Обновить товар",
        "parameters": [
          {
            "type": "string",
            "description": "ID товара",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Обновленные данные товара",
            "name": "product",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Product"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Пометить товар как удаленный (требуется роль manager или director)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Удалить товар",
        "parameters": [
          {
            "type": "string",
            "description": "ID товара",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/products/{id}/restore": {
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Восстановить ранее удаленный товар (требуется роль manager или director)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["products"],
        "summary": "Восстановить удаленный товар",
        "parameters": [
          {
            "type": "string",
            "description": "ID товара",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Произошла ошибка"
        }
      }
    },
    "models.Product": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "status": {
          "type": "string"
        },
        "stock": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}